support static link of SDL2 and pkg-config
Add SDL2_STATIC variable, when set to ON will link the static version of
SDL2 and add any necessary libs and linker flags for linking it
statically from pkg-config. All extra libs and linker flags are added to
SDL2_LIBRARY, so the user does not need to do anything except set that
flag.

For dynamic linking, any extra stuff from pkg-config is also retrieved
and added to SDL2_LIBRARY.

The variable SDL2_DEFINITIONS is also retrieved from
`pkg-config --cflags-only-other`. The user needs to call ADD_DEFINITIONS
for it, the documentation is updated to reflect this.

If pkg-config is not available, the module functions as before with no
errors, but SDL2_STATIC will probably fail.

TODO:

* warn if pkg-config is not found

* add similar code to the other two cmake modules, or factor it out

* support sdl2-config as well (extra parsing)